calculator-app/backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-MySQLdb==2.0.0
mysqlclient==2.2.4
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==22.0.0
